Dialog removed now going to details directly

+++ src/fr/odai/smsdiffusion/FragementCallbacks.java
@@ -3,6 +3,6 @@ package fr.odai.smsdiffusion;
 
 public interface FragementCallbacks {
 	
-	public int getListId();
+	public long getListId();
 
 }
